Update model.py
before:

gt_boxes = KL.Lambda(lambda x: norm_boxes_graph(
    x, K.shape(input_image)[1:3]))(input_gt_boxes)

+++ mrcnn/model.py
@@ -1873,7 +1873,7 @@ class MaskRCNN():
                 shape=[None, 4], name="input_gt_boxes", dtype=tf.float32)
             # Normalize coordinates
             gt_boxes = KL.Lambda(lambda x: norm_boxes_graph(
-                x, K.shape(input_image)[1:3]))(input_gt_boxes)
+    		tf.cast(x, tf.float32), K.shape(input_image)[1:3]))(input_gt_boxes)
             # 3. GT Masks (zero padded)
             # [batch, height, width, MAX_GT_INSTANCES]
             if config.USE_MINI_MASK:
